Feat: client side ics generation and download
